change score to long, so we do not have to worry about overflow

+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/PeerReputation.java
@@ -44,7 +44,7 @@ public class PeerReputation implements Comparable<PeerReputation> {
   private static final int SMALL_ADJUSTMENT = 1;
   private static final int LARGE_ADJUSTMENT = 10;
 
-  private int score = DEFAULT_SCORE;
+  private long score = DEFAULT_SCORE;
 
   public Optional<DisconnectReason> recordRequestTimeout(final int requestCode) {
     final int newTimeoutCount = getOrCreateTimeoutCount(requestCode).incrementAndGet();
@@ -100,6 +100,6 @@ public class PeerReputation implements Comparable<PeerReputation> {
 
   @Override
   public int compareTo(final @Nonnull PeerReputation otherReputation) {
-    return Integer.compare(this.score, otherReputation.score);
+    return Long.compare(this.score, otherReputation.score);
   }
 }
